Clean up auto-loading
