+++ programs/psitest/CMakeLists.txt
@@ -23,9 +23,11 @@ set_target_properties(psitest PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${ROOT_BINARY_
 install(TARGETS psitest RUNTIME COMPONENT Tester)
 
 configure_file(psitest-gdb.py ${ROOT_BINARY_DIR}/psitest-gdb.py COPYONLY)
+configure_file(psibase.py ${ROOT_BINARY_DIR}/share/psibase/python/psibase.py COPYONLY)
 
 include(GNUInstallDirs)
 install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/psitest-gdb.py DESTINATION ${CMAKE_INSTALL_LIBDIR}/debug COMPONENT Tester)
+install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/psibase.py DESTINATION ${CMAKE_INSTALL_DATADIR}/psibase/python COMPONENT Tester)
 # TODO: find out why this is necessary. Without this,
 # psitest, when compiled with gcc, produces
 # "longjmp causes uninitialized stack frame"
